requirements.txt
celery==5.2.7
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
python-decouple==3.7
redis==4.5.1
